+++ src/lib.rs
@@ -21,8 +21,8 @@ mod pypi_marker_env;
 mod pypi_tags;
 
 pub use activation::get_activation_env;
-pub use environment::UpdateLockFileOptions;
 pub use lock_file::load_lock_file;
+pub use lock_file::UpdateLockFileOptions;
 pub use project::{
     manifest::{EnvironmentName, FeatureName},
     DependencyType, Project, SpecType,
